backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
websockets==12.0
pyyaml==6.0.1
trimesh==4.0.5
numpy==1.24.3
psutil==5.9.6
aiofiles==23.2.1

# Optional but recommended
python-dotenv==1.0.0
rich==13.7.0